Now we use cost price as price on purchase documents.
